+++ doc/changelog.rst
@@ -11,6 +11,8 @@ Released on <unreleased>
 
 This release is a preparation for a stable 1.0 release.
 
+- Fixed comparison operators when working with empty or unbounded ranges. They
+  would previously raise exceptions.
 - Renamed classes to match :pep:`8#class-names` conventions. This does not apply
   to classes that works on built-in that does not follow :pep:`8#class-names`.
 - Optimized :meth:`~spans.types.Range.overlap`, 1.3x speedup.